pyproject.toml
[tool.poetry]
name = "to_do_list"
version = "0.1.0"
description = ""
authors = ["Luiz Antonio Lazoti <luizlazoti@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10.12"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.30"
pytz = "^2024.1"
fastapi = "^0.111.0"
alembic = "^1.13.1"
requests = "^2.32.3"
jinja2 = "^3.1.4"
psycopg2-binary = "^2.9.9"
ruff = "^0.6.9"
uvicorn = "^0.31.1"
gunicorn = "^23.0.0"
faker = "^30.3.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
